Moving boavizta to directory
