+++ src/lib/boavizta/index.test.ts
@@ -1,9 +1,9 @@
 import {describe, expect, jest, test} from '@jest/globals';
-import {BoaviztaCloudImpactModel, BoaviztaCpuImpactModel} from './boavizta';
+import {BoaviztaCloudImpactModel, BoaviztaCpuImpactModel} from './index';
 import axios from 'axios';
-import * as PROVIDERS from '../__mocks__/boavizta/providers.json';
-import * as COUNTRIES from '../__mocks__/boavizta/countries.json';
-import * as INSTANCE_TYPES from '../__mocks__/boavizta/instance_types.json';
+import * as PROVIDERS from '../../__mocks__/boavizta/providers.json';
+import * as COUNTRIES from '../../__mocks__/boavizta/countries.json';
+import * as INSTANCE_TYPES from '../../__mocks__/boavizta/instance_types.json';
 
 jest.mock('axios');
 const mockAxios = axios as jest.Mocked<typeof axios>;
